app/requirements.txt
Django>=3.0,<4.0
PyYAML>=5.4.1
djangorestframework>=3.12.4
django-filter>=2.4.0
django-cors-headers>=3.7.0
markdown>=3.3.4
uWSGI>=2.0.19.1
channels>=3.0.4
channels-redis>=3.3.0
libsass
django-compressor
django-sass-processor
